backend/api_service/requirements.txt

fastapi==0.115.5
uvicorn[standard]==0.30.6
pydantic==2.9.2
google-cloud-firestore==2.19.0
google-cloud-pubsub==2.27.0
google-cloud-storage==2.18.0
firebase-admin==6.5.0
